nervesparks/requirements.txt
streamlit
PyMuPDF
pytesseract
pillow
sentence-transformers
chromadb
openai